Ensure all buffers are registered in Zeta CLI
